Foldy Lax Model

Background

To test the dictionary learning algorithms, DictionaryLearning.jl provides utilities for generating Green's matrices for the Foldy-Lax model of scattered wave-fields. In this section we give an overview of this model and what is implemented in this package.

For some $a,b \in \mathbb{R}^3$ define $f(a,b) : \mathbb{C}^p \to \mathbb{C}^p$ to be the linear transformation such that

\[f(a,b)e_s = \frac{\exp\left( \frac{2\pi f_s}{c_0} i\left\lVert a - b\right\rVert_2\right)}{4\pi \left\lVert a - b\right\rVert_2}e_s\]

This is the free-space propogator. $f(a, b_i)$ takes a signal emitted by a source at $b$ to the signal recieved by a reciever at $a$.

Let $r_1, \cdots, r_n \in \mathbb{R}^3$ be the coordinates of our recievers. Let $\xi_1, \cdots, \xi_m\in \mathbb{R}^3$ be the scatterers, $\hat\xi_1, \cdots, \hat\xi_m \in \mathbb{C}^p$ their signal, and $0 \leq \tau_1, \cdots, \tau_m \leq 1$ be their scattering amplitudes. Let $z_1, \cdots, z_k \in \mathbb{R}^3$ be the coordinates of the sources, and $\hat z_1, \cdots \hat z_k \in \mathbb{C}^p$ their signal.

\[\hat{r}_i = \sum_{j=1}^k f(r_i, z_j)\hat z_j + \sum_{j=1}^m f(r_i, \xi_j)\tau_j \hat \xi_j \qquad i = 1, \cdots, n \tag{1}\]

\[\hat{\xi}_i = \sum_{j=1}^k f(\xi_i, z_j)\hat z_j + \sum_{\substack{j=1\\ j \neq i}}^m f(\xi_i, \xi_j) \tau_j \hat \xi_j \qquad i = 1, \cdots, m \tag{2}\]

Given $\hat z_i$ as well as coordinates $r_i$, $z_i$, and $\xi_i$ in general positions, equations (1) and (2) constitude a determined system of linear equations.

For $s = 1, \cdots, p$ define $\mathcal{G}_s : \mathbb{C}^{k} \to \mathbb{C}^{m}$ to be the linear transformation such that

\[\mathcal{G}_s \begin{bmatrix} e_s^T \hat z_1\\ \vdots\\ e_s^T \hat z_k \end{bmatrix} = \begin{bmatrix} e_s^T \hat r_1\\ \vdots\\ e_s^T \hat r_m \end{bmatrix}.\]

Then, define $\mathcal{G} = \mathcal{G} = \begin{bmatrix}\mathcal{G}_1^T, \cdots, \mathcal{G}_s^T\end{bmatrix}^T$. This is the dictionary we would like to learn. Forming $\mathcal{G}$ efficiently is the goal of the of the functions below. It suffices to construct the $\mathcal{G}_s$ independently and in parallel. With this in mind, define

\[\hat z^s = \begin{bmatrix}\hat z_1, \cdots, \hat z_k\end{bmatrix}^T e_s\qquad\text{and}\qquad \hat r^s = \begin{bmatrix}\hat r_1, \cdots, \hat r_m\end{bmatrix}^T e_s\]

Define $M^{[\xi z]}_s \in M_\mathbb{C}(m, k),$ $M^{[rz]}_s \in M_\mathbb{C}(n,k)$, $M^{[\xi \xi]}_s \in M_{\mathbb{C}}(m, m),$ and $M^{[r \xi]}_s \in M_\mathbb{C}(n, m)$ so that

\[\begin{align*} e^T_i M^{[\xi z]}_s e_j &= e_s^T f(\xi_i, z_j)\\ e^T_i M^{[r z]}_s e_j &= e_s^T f(r_i, z_j)\\ e^T_i M^{[\xi \xi]}_s e_j &= \begin{cases} e_s^T f(\xi_i, \xi_j)\tau_j & i \neq j\\ -1 & i = j \end{cases}\\ e^T_i M^{[r \xi]}_s e_j &= e_s^T f(r_i, \xi_j) \tau_j \end{align*}\]

Solving (2) gives the excited waves in terms of the source waves:

\[\hat \xi^s = - \left(M^{[\xi \xi]}_s\right)^{-1} M^{[\xi z]}_s \hat z^s. \tag{3}\]

Similarly, we may rewrite (1) as

\[\hat r^s = M^{[rz]}_s \hat z_s + M^{[r \xi]}_s \hat\xi^s.\]

Substituting into (3) lends

\[\hat r^s = \left[M^{[rz]}_s - M^{[r \xi]}_s \left(M^{[\xi \xi]}_s\right)^{-1} M^{[\xi z]}_s\right]\hat z^s. \tag{4}\]

The quantity inside the brackets is $\mathcal{G}_s$.

Functions

See the example in /examples/foldy-lax/foldy-lax-example.ipynb for how to use the following.

DictionaryLearning.FoldyLaxStruct — Type

struct FoldyLaxStruct{T <: Real}

Properties

G::StridedMatrix{T}: Array of size (p*n, k) containing the matrix $\mathcal{G}$ described above.

Stores parameters for and results of computation. Importantly, after calling foldylax!, the property G will contain the matrix $\mathcal{G}$ described above.

function FoldyLaxStruct(src, sct, rec, frq, τ) where T

Arguments

src::StridedMatrix{T}: Array of size (d, k) where each column is the coordinate of a source.

sct::StridedMatrix{T}: Array of size (d, m) where each column is the coordinate of a scatterer

rec::StridedMatrix{T}: Array of size (d, n) where each column is the coordinate of a reciever

frq::StridedVector{T}: Vector of length p where each element specifying the frequencies

τ::StridedMatrix{T}: Vector of length m where each element stores the scattering strength of the corresponding scatterer

DictionaryLearning.FoldyLaxWorkStruct — Type

struct FoldyLaxWorkStruct{T <: Real}

Stores scratchspace required by foldylax!. This can be reused by subsequency calls foldylax! so long the geometry specified in the corresponding FoldyLaxStruct{T} has the same dimensions.

function FoldyLaxWorkStruct(fls)

Create an instance of FoldyLaxWorkStruct that is compatible with fls

Arguments

fls::FoldyLaxStruct{T}

DictionaryLearning.foldylax! — Function

function foldylax!(fls, flws) where T

Overwrites fls.G with the matrix $\mathcal{G}$ described above.

Arguments

fls::FoldyLaxStruct{T}

flws::FoldyLaxWorkStruct{T}