Utilities

DictionaryLearning.gen_sparse_samples — Function

function gen_sparse_samples(T, k, n, s)

Construct a sparse matrix X of size (k, n) such that each column contains s non-zero elements. The values of these non-zero elements is chosen randomly on the real or complex sphere as specified by T.

Arguments

T::Union{Type{Float32}, Type{Float64}, Type{ComplexF32}, Type{ComplexF64}}: Element type of X

k::Int: First dimension of X

n::Int: Second dimension of X

s::Int: Sparsity of each column of X

DictionaryLearning.align_dict — Function

function align_dict(A, B)

Computes the permutation of columns of B and unit-magnitude scaling of those columns so that the described matrix $\tilde B$ minimizes $\lVert A - \tilde B\rVert_F$. It is assumed that the columns of A and B are normalized.

Arguments

A::StridedMatrix{T}

B::StridedMatrix{T}

Returns

perm::Vector{Int}: Vector of length k describing a permutation of the columns of B.

rot::Vector{T}: Vector of length k describing how to scale the columns of B.

See the implementation of align_dict! for how to compute $\tilde B$ from $B$.

DictionaryLearning.align_dict! — Function

function align_dict!(A, B)

Same as align_dict except B is transformed in-place as well.

DictionaryLearning.subdist — Function

function subdist(A, B, nrm=:opnorm)

Computes the subspace distance between the column-space of A and B under the norm specified by :opnorm. It is assumed that the columns of A and B form an orthonormal basis.

Arguments

A::StridedMatrix

B::StridedMatrix

nrm::Symbol: Either :opnorm or :fnorm for operator norm and Frobenious norms respectivelly.

DictionaryLearning.max_cov! — Function

function max_cov!(M, work)

Compute the maximum covariance between columns distinct pairs of columns of M.

Arguments

M::StridedMatrix: A matrix of size (n, m)

work::AbstractArray: Scratch space of length at least n. More scratch space will usually improve performance, particularly on GPUs.

DictionaryLearning.get_free_mem — Function

function get_free_mem(back)

Returns the amount of free memory on the device specified by back. Useful for allocating work space.

Arguments

back::StridedMatrix